+++ src/UI/Components/LVGL/LvObj.h
@@ -233,7 +233,7 @@ namespace UI
 
 		void moveToFront();
 		void moveToBack();
-		void moveToIndex(size_t index);
+		void moveToIndex(int32_t index);
 		void clearChildren();
 
 		void setVisible(bool display, bool move_to_front = false);
